+++ packages/llms-e2e-test-utils/src/wp-version-compare.js
@@ -1,0 +1,16 @@
+import { cmp, coerce } from 'semver';
+
+import { getWPVersion } from './get-wp-version';
+
+/**
+ * Run a version compare against the currently tested version of WordPress.
+ *
+ * @since [version]
+ *
+ * @param {string} version    A version string.
+ * @param {String} comparator A comparison string, eg ">=" or "<", etc...
+ * @return {boolean} Comparison result.
+ */
+export function wpVersionCompare( version, comparator = '>=' ) {
+	return cmp( coerce( version ), comparator, coerce( getWPVersion() ) );
+}