+++ packages/llms-e2e-test-utils/CHANGELOG.md
@@ -1,10 +1,17 @@
 LifterLMS E2E Test Utils Changelog
 ==================================
 
+v3.2.0 - 2021-01-31
+-------------------
+
++ Added: New function `wpVersionCompare()` used to run version comparisons against the currently tested version of WordPress.
++ Fixed: Tests failing when running `runSetupWizard()` on WordPress >= 5.9.
+
+
 v3.1.0 - 2021-12-07
 -------------------
 
-+ Added new functions `highlightCode()` and `setCheckboxSetting()`.
++ Added new functions `highlightNode()` and `setCheckboxSetting()`.
 
 
 v3.0.0 - 2021-11-05